First import of autolevel correction class (adapted from the Gimp2.0). Added menu entry in ImageDitor for AutoLevel correction.
Renchi : the autolevel correction code is completed excepted for the source
image data manipulation method, because i haven't found this method in the Gimp
(the big bazard !!!). The 'Auto Level' option don't change the image, just
calculate the level and the lut. TODO : apply the lut to the image (how to do that ?)

CCMAIL: renchi@pooh.tam.uiuc.edu

svn path=/trunk/kdeextragear-3/digikam/; revision=333784