Digikam Image Editor free rotation tool : added a tools to guide the user for good image rotation adjustment.
To Digikam users : move the mouse cursor on the free rotation preview and horizontal and vertical dashed lines will be draw. Place this lines to
an important place in the image (the sea or a building for example) and press the left mouse button. The lines position will be frozen like a reference. Now adjust the free angle to a correct position accordinly with the doashed lines.

CCMAIL: digikam-users@lists.sourceforge.net

svn path=/trunk/kdeextragear-3/digikam/; revision=363393