New digiKam image editor plugin : Blur FX.
With this plugin you can apply blurring special effects to a photograph.
Current effects are available :

- Zoom Blur:  blurs the image along radial lines starting from a specified
center point. This simulates the blur of a zooming camera.

- Radial Blur: blurs the image by rotating the pixels around the specified
center point. This simulates the blur of a rotating camera.

- Far Blur: blurs the image by using far pixels. This simulates the blur of an
unfocalized camera lens.

- Motion Blur: blurs the image by moving the pixels horizontally. This simulates
the blur of a moving camera.

- Skake Blur: blurs the image by skaking randomly the pixels. This simulates the blur of a random moving camera.

- Softner Blur: blurs the image softly in dark tones and hardly in light tones.
This gives images a dreamy and glossy soft focus effect. It's ideal for
creating ro:mantic portraits, glamour photographs, or giving images a warm
and subtle glow.

- Frost Glass: blurs the image by randomly disperse light coming through a
frosted glass.

A screenshot is available at this url :

http://extragear.kde.org/apps/digikamimageplugins/blurfx.png

CCMAIL: digikam-devel@lists.sourceforge.net

svn path=/trunk/kdeextragear-3/digikamimageplugins/; revision=393822