CCBUGS: 105303
give a more detailed info if user uses symlinks to external files/folders
in the album library path and then tries to move it. I'm not going to work
on this in the trunk (0.7.x series). The file copy/move kioslave will undergo 
a rewrite in the experimental branch to work well with kio::jobs and if 
necessary will be backported


svn path=/trunk/extragear/graphics/digikam/; revision=412383