After I've played some time with the new AlbumFolderView / TagFolderView I realized, that it is sometimes impossible to add a new album to the root album or dragging an album to the root.
I've added a new root folder to the two views, which are displayed red, 
to show their different meaning. Additionally they are displayed without
rootDecorations.

You can rmb click to that root album to add a new child to the root or 
drag another album on it. 

Digikam doesn't allow images to be stored in the root folder.

Todo: disable the some parts of the import menu, when the root album
is selected.

CCMAIL: digikam-devel@kde.org


svn path=/trunk/extragear/graphics/digikam/; revision=438029