Don't using KFilePreview thumbnails cache to display properly the image modifications on showfoto thumbbar. Please, give me a feedback...
Nota: using KIO:KfilePreview isn't the better way to render thumbnails. This method don't work properly with any 
TIFF and PNG files generated with imlib2 (especially for example with hight PNG compression level). 
Perhaps the better way is to use the digikamthumbnails kioslave when it's available (for example in a both
installation of showfoto and digikam in the same computer).

CCBUG: 111873


svn path=/trunk/extragear/graphics/digikam/; revision=456411