digikam 0.9.0 unstable (trunk) : kioslave::digikamthumbnail is now based on DImg instead imlib2. now TIFF and PPM 16 bits images are thumbnailized properly.
NOTA: to improve digikamthumbnail  kiosilave, we need to add methods in DImg to get embedded thumnail in image if available (PNG, TIFF, JPEG, etc...) instead to load the entire image. Sure, image must have been saved including an embedded thumnail and we need to add this feature in image editor when image are modified and saved. 

svn path=/trunk/extragear/graphics/digikam/; revision=483286