digikam from trunk : simplify and fix dirty image flag if current image has changed. Removed m_dirtyImage flag witch cannot be use properly with any flag event from canvas (race condition). Sometimes IE/showfoto said if user want saving current image more than one time...
To the team : I have checked indeep all cases in IE and showfoto when current image is changed and with all user interactions :
- changing current image;
- save and saveas;
- removing current image;
- Loading images/folder (showfoto)

All seam to work fine now. Please, check if all working fine for you. If yes, this commit must be backported to stable...

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=496058