digikam from trunk: Port "Insert text" plugin to DImg
This one was a bit difficult, because we have no pixel access to font rendering in Qt3.
So we must still use a QPixmap to draw the (antialiased) text, which means we will
lose color resolution in the area where the text is drawn for 16 bit images.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikamimageplugins/; revision=522215