digikam from trunk : do not include file name extension during file renaming operation from users.
Note : with the old implementation, when the the filename extension is changed by user, we will check if this one is in the mimetype setup. if no, it's added, bu _always_ like an image mimetype !!! Imaging if you put a sound or a video mimetype in the image mimetype list : image editor will trying to open it... 

The mimetype registration with renaming is now removed. 

BUG: 117375
CCBUGS: 113801

svn path=/trunk/extragear/graphics/digikam/; revision=566834