digikam from trunk : main digikam/digikam folder compile with -nofinal option from .configure script now. Before, enable-final have been disable for this folder.
digikam/digikam/Makefile.am is now fixed and all compile fine in this mode. The files order to compile libdigikam.la is very important in this mode. We need to take a care about this point in the future.

Please report me if you have any problems on your Linux dist when you compile packages. Thanks in advance

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=609382