digikam from trunk : Continue 0.9.1 implementation !
Second stage : adding a new Welcome page on the root album of your pictures collection. The page is inspired of Kmail or Konqueror Welcome page. On root album there is nothing to see/to do. This is the right way to show a resume of digiKam features dedicaced to new users.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=614735