Introducing the class MetadataHub. It is supposed to be the backend for - Support for multiple selection - unify code to write metadata
The class can be used to
- read from ImageInfo objects
- read from DMetadata objects
- merge the information
- change the contained data fields (date, comment, rating, tags)
- write to Metadata
- write to ImageInfo
- retrieve the value of the data fields

Many datasets can be merged. If the sets contain different data,
the status of the field (date, comment, rating, single tags) will be "disjoint",
the status will be "available" if all contained the same data.

Note that the code is not yet in use or tested.


svn path=/trunk/extragear/graphics/digikam/; revision=621841