digikam from trunk : forcing focus to Comments Edit Widget from Comments & Tags sidebar is not the right solution. Focus stuff  is a complicated problem witch must be study indeep.
In fact the folder view and album icon view take the focus with the mouse. The previous implementation cannot work with that. Also, alots of menu shorcuts are broken if we force Comments edit widget to take focus.

My viewpoint is to implement a shortcut to toggle gui to captionning mode. It a better/simple solution. Please give me your viewpoint.

CCBUGS: 131632, 131743

BUG: 140412

svn path=/trunk/extragear/graphics/digikam/; revision=628812