Disable the use of the DImg raw loader when loading as DImg from digikamthumbnail ioslave.
We are using KProcess for the raw loader and make two assumptions:
 - there is an event loop (not for ioslaves)
 - the loader is not called from the event loop thread (currently no use case in digikam)

The DImg raw loader is not needed - DcrawIface is called before.
So the problem of the loader entering an endless loop (because no events from KProcess
are delivered) arises only if a file is identified as RAW, but dcraw does not
support it.


svn path=/trunk/extragear/graphics/digikam/; revision=631905