digiKam from trunk : Image Editor : severe bug fix in canvas :
With implementations <= 0.9.1, the canvas doesn't work properly using high zoom level (> 500). The real image <==> canvas pixmap interface do not compute properlly the tile area values. This provide a lots of  "tile" artefacts over the image. Look the example here:

http://digikam3rdparty.free.fr/Screenshots/editorhighzoomartefact.png

Note than these "tile" artifacts are not the real tiles of canvas and of course not the JPEG artifact (the image is PNG in this screenshot)

The new implementation fix this problem to handle properly the areas to 
use from the source image to generate the canvas pixmap tiles.  

Note to users : i need your feedback to check if all work fine on your computer (to be sure if my patch do not give side effects (:=)))). Thanks in advance...

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=649987