digiKam from trunk : Light Table : Add "Navigate by Pair" option.
In this mode drag & drop is disable between thumbbar and left/right panels.
The options to set amnually the item on Left or Right panel are also disable (main menu and pop-up menu)

The light table will set automaticly the left item with the current item selected on thumbbar and the right item 
is the next item on thumbbar.

CCBUGS: 145236

svn path=/trunk/extragear/graphics/digikam/; revision=663963