digiKam from trunk : Light Table preview : fix tile side effect with high zoom level. This is a first approach how to solv ethis problem. It's not optimum but work fine.
What the problem exactly ? Well it's simple : we scaling a part of full image to render a preview unsing a tile matrix. The part of ful image is scaled with take a care of 
image data outside the region. With a high zoom level the scale algorithm cannot compute properlly the border of the tile and a side effect appear.

CCBUGS: 145198

svn path=/trunk/extragear/graphics/digikam/; revision=667096