Add the first real schema updating.
All previous versions do not contain code to detect database schemas,
this means they cannot find out that a database is in a future schema
that they cannot read. For this reason, first step is to copy the digikam3.db
file to digikam4.db.
Further schema changes are not yet implemented, but from now on they can be done
on this file.
I am open for discussions on compatibility issues in the future, this is only a first implementation.

svn path=/branches/work/digikam-dbconsolidation/; revision=669944