Add a pair of classes to implement our style of sharing data, the good old "explicit caching" with some extra methods for my special needs.
Large parts of the code are taken from qshareddata.h (hence the slightly different copyright header - GPL v2 only)
This code uses QAtomic methods, but it is from a _public header_ which means that
the API used there - even if not officially documented - will not be subject to change (to keep BC).

The class still needs to be tested - it only compiles.

svn path=/trunk/extragear/graphics/digikam/; revision=678618