Switch to multithreaded thumbnail loading: AlbumIconView uses ThumbnailLoadThread now.
Previously, we used to load thumbnails with an ioslave and transfer the data over shared memory.
This overhead can be avoided now.

The new classes ThumbnailLoadThread and ThumbnailCreator have the functionality that was 
contained in the digikamthumbnail ioslave and the classes ThumbnailJob and PixmapManager.
ThumbnailJob/ioslave is still used in other parts of digikam. PixmapManager is now unused.

Subjectively, thumbnail loading feels faster in certain situations for me.
I have no measurements because the experimental setting would be difficult.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=698276