Move filtering by name from listing to scanning.
Previously, all files were taken in the database, and only after listing from the db,
files were filtered by the file suffixes.
With this change, as discussed on the devel list quite some time ago,
files are filtered by suffix before they are taken up in the database,
and filtering after listing is not necessary.

This accommodates better to the view that all entries in the db are part of the collection,
and it solves some bugs when SQL features are used to get statistics from db.

TODO:
The situation in the config dialog needs to be cleared. At least there must be a big fat
warning if users try to remove "*.jpg" and the like, as they would lose their tags.
Second, after a change, a rescan is needed.
We should anyway think about the format list, when raw formats are added, old users may not see them.
There could be a "use default values" check - and only with this unchecked, the formats would be configurable.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=700513