digiKam from trunk (KDE4) :
Marcel, 

with future CameraGui implementation based on KXmlGuiWindow, the Naviguate bar embeded in imageproperties 
and metadata sidebar tab will be not used.
This naviguate bar is obsolete now. We will use a status naviguate bar embeded in StatusBar instead everywhere.

This commit disable it in GUI, but the better way is to simplify the implementation in the future 
to remove it definitivly.

CCMAIL: marcel.wiesweg@gmx.de


svn path=/trunk/extragear/graphics/digikam/; revision=706087