At the time it was a nice idea to write the absolute path and the status
of album roots dynamically into the database so that they were available
from SQL.
But now I see that this is not the way to go. It does not allow sharing DBs
(not that we support that officially, or say that it works, but people will do
it once we support MySQL, and this would break it), and it does not work
with read-only DBs.
So now, we always have the album root id and use CollectionManager then
which has all the necessary information.
Adapt a lot of SQL.

svn path=/trunk/extragear/graphics/digikam/; revision=726439