digiKam from trunk (KDE4): following Marcel tips by IRC, We use now a dedicaced Multithreaded Thumbnail Loader instance from digiKam core for all kipi-plugins we need thumb. We never use default KDE thumbnails loader with all kipi-plugins running under digiKam. Why ? : because KDE thumnails loader is slow, do not work properlly with JPEG and PNG, do not support RAW pictures, do not rotate properlly thumbs accordinly to Exif or Xmp metadata, etc...
It's perfect... To test it, start SendImages kipi-plugin for ex. and add new pictures to e-mail on the list.
Look how is render the pictures thumb on the left of open file dialog (Standard KDE + new preview widget). 
This rock! digiKam and kipi-plugins because more homogenous now.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=740390