To uniquely identify the storage media of an album root,
the UUID is a very good identifier, but it is not always available.
There need to be other means of identification for the storage of an album root.
Change the format of the album root identifier to a KUrl based string
which allows easy storage of a variety of identifiers:
- UUID
- label
- hash of the file names in the top level directory (CD/DVD)
- mount path
- path

Add private methods to create such identifiers, match Solid volumes to a given identifier,
and find a volume for a given path. Adapt the code to the new situation.

Add a method to return an user-visible information string on how a given path can be detected.
This can inform the user e.g. that a CD is identified by its label, or a removable
USB disk can be uniquely identified by its UUID.

This commit invalidates existing album roots; Either delete and readd them,
or edit the AlbumRoots table field "identifier" by prepending "volumeid:?uuid=".

svn path=/trunk/extragear/graphics/digikam/; revision=741554