New Search Interface:
SearchTabHeader is the widget at the top of the search sidebar.
It provides a directly accessible line edit to start a quick search, and a button
to start an advanced search.
In the lower part, the current search can be saved as a stored search,
or if a stored search is selected, its parameters can be edited.

Note:
The widget currently consists of two group boxes. This may not be
the best solution, but after some experiments with separator line, flat group
boxes, QFrames etc. this was the most acceptable solution so far.
Note:
Advanced search is still under development; you will see that it will
show an UI, but not carry out a search.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=792262