Fix a bug in the proxyclicklineedit: Must be mouse release, not mouse press.
Add a class ListViewComboBox. Any QComboBox has obviously a QListView as default view
(with minor additions - see qcombobox_p.h in Qt source, but it must be safe to use any view),
but for StayPoppedUpComboBox we need a public viewportEvent() which requires subclassing
and replacing the view. Now, ListViewComboBox is a default implementation of StayPoppedUpComboBox

svn path=/trunk/extragear/graphics/digikam/; revision=799178