Add geodetic tools:
GeodeticCalculator allows to
 - compute the distance between two points on earth
 - compute the second point, given a starting point, a distance and an azimuth

The algorithm is assuming the earth in an ellipsoid, and this is the best
assumption available to make such computations.

This code is taken from the geotools project, LGPL 2.1, now relicensed
to GPL 2 or later (If equivalent code is provided by Marble,
we can use that one. I did not find any.)

svn path=/trunk/extragear/graphics/digikam/; revision=806826