digiKam from trunk (KDE4) : I'm happy to said than a new powerfull Search tool is available in digiKam : Searches over a map !
This tool use Marble widget from trunk with a patch from me to perform a rectangle selection on the canvas. Thepatch is available at this url:

http://digikam3rdparty.free.fr/misc.tarballs/marbleselection.patch

When you make a selection, digiKam query the database to find all pictures which have been taken into this area.

There is a screenshot of this tool in action at this url:

http://digikam3rdparty.free.fr/Screenshots/NewSearchTools/mapsearchtool.png

Several improvments need to be done in marble to have a more suitable selection tool, but at least it work fine as well.

CCMAIL: marcel.wiesweg@gmx.de
CCMAIL: digikam-users@kde.org
CCMAIL: tackat@kde.org

BUGS: 153070



svn path=/trunk/extragear/graphics/digikam/; revision=816774