Scan for images in the root directory of an album root as well.
Historically, digikam had one album root and all subdirectories of this root
were used as albums. Images in the root directory itself were not included in
the collection (after all, there was no album for them).
Now that we can have multiple album roots, and the album roots are represented
with an album in the album view, this exclusion is no longer necessary.
This is important in particular for users who want to add one single directory
containing images, but no subdirs, as an extra album root. Previously this would
have resulted in one empty album, now it's still one album, but it's filled.

svn path=/trunk/extragear/graphics/digikam/; revision=818009