Avoid deadlock:
Do not hold a mutex when emitting a blocking queued signal.
In this case, ensure that DatabaseAccess is released, at least temporarily.
Otherwise there will be a deadlock when emitting the synchronous signal to
retrieve Solid information, and the main thread is currently handling a signal
that wants to acquire DatabaseAccess.
Adds a class that unlocks the recursive mutex on creation, and reqacuires on destruction.
This should be used with care as it undermines assumptions that structures
cannot change while the lock is held. But you should not hold a mutex when calling
methods with side effect (we do this here in at least one situtation), so
I happily break assumptions (and there are no such assumptions currently).

Thanks to Thomas McGuire for providing backtraces.

CCBUG:165348

svn path=/trunk/extragear/graphics/digikam/; revision=826917