Introduce two more default thread objects. We use default threads everywhere,
it's better to have separate threads e.g. in the icon view and in the thumbbar,
but it's ok to share threads between all thumbbars.

Add versions of the find methods that override the default thumbnail size.
Using a global size per thread looks better API-wise but does not carry
well with the way we share the threads. Need to look at some places still
if they properly set the size.

svn path=/trunk/extragear/graphics/digikam/; revision=828363