By caching all thumbnails for all items in a thumbbar we create a number of problems,
especially if the album (all of whose items are in the thumbbar) has more than 100 images:
- the global cache policy is circumvented
- implicitly cached objects, if another instance is kept, will not be removed
  from the QCache. This can lead to a situation where the thumbbar items
  block the whole cache, and new items cannot be stored. The iconview
  depends on at least all visible items available from the cache
- smooth scaling at painting is slow, caching a pixmap of the right size
  is much faster for drawing
- with even larger albums, think of a search album, memory usage can be huge

This commit brings the thumbbar to the same behavior used by icon view,
taking pixmaps from the cache or triggering to load them.
We can implement more preloading (the next x items etc.) if we need it.

svn path=/trunk/extragear/graphics/digikam/; revision=828365