Add special handling for the case the the requested tile size is larger
than the allowed maximum thumbnail size.
In this case, we scale up the image, which results in bad-quality blurred
versions pretending to be larger than the contained information allows.

I think we should set the thumbbar's maximum size so that the user
cannot drag it larger than for the 256 thumbnails.

svn path=/trunk/extragear/graphics/digikam/; revision=828368