re-factorize histogramwidget in all editor tools and sidebars
Until now the histogram widget in the editor tools and the sidebar has been implemented by copy & paste.
Changes to that code had to be made in all of the 13 editor tools and the
sidebars that use a histogram.
Also the slots slotChannelChanged() and slotScaleChanged() where implemented in every single
editor tool, although they use a common code base most of the time (only channel mixer, adjustcurves
and adjustlevels use slightly different slots).
So it is time to factorize this code.

This patch will move all histogram related code into a new widget:
HistogramBox.
This widget holds a histogram widget as well as all the control widgets for it,
like the scale buttons, the channel selector and the optional gradient widget.

To use the histogrambox in editor tools, a new toolcode is introduced to enable
the histogram. To set its mode, a third parameter 'histogramType' is added to
the ctor of EditorToolSettings. The different modes for the HistogramBox are
defined in the enum HistogramType.

HistogramBox defines the following methods:

----------------------------------------------------------------------

histogram() :
    returns the histogram widget

setGradientVisible(bool) :
    show / hide the color gradient widget beneath the histogram widget.
    Some tools like adjustcurves and adjustlevels don't need that widget so we need a way to
    disable it

setGradientColors(const QColor &from, const QColor &to) :
    set the color range for the gradient widget.

channel() :
    returns the selected channel (Luminosity, RGB etc). Since a QVariant is used in the channel combobox,
    we need to convert the ID of the combobox item to the actual channel value. This method handles the
    conversion.

setChannel(int channel) :
    set the current histogram channel. Due to the use of the QVariant we need to convert the channel id properly.
    This method handles the conversion.

setColorsEnabled(bool) :
    if 'Colors' channel can be used in the editor tool, we need to provide a method to enable / disable this
    combobox.

setColors(int color) :
    set the current ColorsChannel

scale() :
    get the current selected histogram scale (linear, logarithmic)

setScale(int scale) :
    set the current histogram scale

----------------------------------------------------------------------

Andi

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=866299