added Inkscape project files for dpopmenu side image
I think it is much easier to change text that has been rotated in an Inkscape project, because you can edit it directly. In a Gimp project it's not possible.
Also the output is much more crisp, the font is not so washed out.
I guess we should remove the Gimp project files...

In addition we could merge both Inkscape project files and work with layers. Since we have same version numbers for digiKam and showFoto, we could put each app name on its own layer and just export the appropriate one when saving the new image.

Andi

svn path=/trunk/extragear/graphics/digikam/; revision=877379