expand and collapse folder view when using quick filter text search
When searching for albums and tags with the quick filter text search, the
folder view has not been expanded when an item matched the search criteria.

For this to work, we have to use setOpen() on every matched item. In most of
the cases the whole album or tag tree will be expanded after you reset the
filter again, which will be very annoying when you have a lot of tags or albums.

This commit will introduce a new FolderView method collapseView(),
that will collapse the tree when the search criteria is removed. Right now it has two modes:

* RestoreCurrentAlbum (default):
  Collapse the folder view and re-open the current viewed album.
  In this mode, all root items are collapsed, then the one containing
  the currently selected album is expand again.
  This mode will make sure that the selected album is visible in the folder tree by scrolling the
  view to the selected item.

* OmitRoot:
  Collapse the folder view but omit the root item.
  In this mode all items in the folder view are collapsed,
  and the first root item is expanded again (My Tags / My Albums etc)

CCBUG:167873

svn path=/trunk/extragear/graphics/digikam/; revision=880036