A custom tree view, model and delegate for the collections setup. - The view is a QTreeView - The delegate inherits KWidgetItemDelegate / Goya framework to provide   widgets inside the view. Rendering is done through an object of QStyledItemDelegate - The model is custom. It keeps a list of collections, allows editing and does   all the work of checking and applying the changes to CollectionManager
Collections are sorted in the three categories local, removable and network.
An "Add collection" button is provided for all three. We can autodetect local and
removable storage media, but we cannot autodetect that a path is a network share
currently, so this solves the problem of providing an extra button for network collections.
When adding a collection the path dialog pops up directly, then the path is checked,
and in a second dialog the user can adjust the name of the collection and is informed
about the status of the storage.
This reintroduces modal dialogs that we did not need with the previous implementation,
but 1) the directory dialog was needed usually anyway and 2) we can display much more
and flexible information.

Label editing is done in-place with standard delegate editor means.
Deleting is done with a small icon button for each collection.

svn path=/trunk/extragear/graphics/digikam/; revision=884632