We can't use KStandardAction::Find for three different actions here, it will mess up the menus and toolbars. Also we need to call KStandardAction::find(), not ::Find() to customize the standard action, otherwise we will have duplicate menu entries and the action will appear in the toolbar.
Gilles, Marcel,

simple search is not focusing / switching to the "search tab" on the left automatically,
so this action is not very useful if the "advanced search" tab is
closed or not displayed at all. It only works if the tab is already
opened.
I think we need to focus the tab automatically, like when using the "Find
similar" action.

CCBUG:178382

svn path=/trunk/extragear/graphics/digikam/; revision=903155