Fix saveViewState() for folderviews with CheckListItems, like TAlbumListView and TagFilterView.
With current implementation, we are not able to have one common
loadViewState() and saveViewState() method, so we need to re-implement
them in TAlbumListView and TagFilterView. This is not nice because we
have duplicate code now, but since we are going to switch to ModelView
concept next release, it would be a waste of time to create a new common
interface or abstract class now.

We can not call virtual methods in constructors / destructors, so every
view inherited from FolderView needs to call saveViewState() in its own
destructor instead.

BUG:148166

svn path=/trunk/extragear/graphics/digikam/; revision=919579