Adding the lines that switch on using a thumbnail database - not enabled per default:
NOTE:
You need to remove the comment signs around albummanager.cpp, lines 685-687, to test the thumbnail database.

For simplicity, it is always created in the directory of the main database and named "thumbnails-digikam.db".
It is using the PGF format to store thumbnails. I am aware there are suggestions to use JPEG or Djvu.

Only real world testing can show if this database is a good thing or not.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=980402