Removing this code again, it will not make things faster.
I did a lot of I/O performance testing now and at least with our sqlite3
solution, we really have reached the limit, it is just not performing
well on bigger collections.

My thumbsDB is 970MB big, and when the disk cache has not been already
filled, querying the database is painfully slow.
This is not digiKam's fault, it can be easily reproduced in the sqlite3
command line tool.

as root:
sync
echo 1 > /proc/sys/vm/drop_caches

Now your disk cache is cleared, as if the system just started up.
Run a simple query on a bigger thumbsDB like:

select count(*) from thumbnails where type = 0;

Even though this query delivers no results, it takes up to 20 seconds to
completely parse the table.

When used in digiKam, the application is not responding at all during
this operation, you get the impression that digiKam just hung up.

I have a large RAM here (4GB), but I don't know what happens if you run
digiKam 1.0.0 on an average system with 512MB or 1GB RAM.

For me the only solution is to go with embedded MySQL or even a real
MySQL server solution.
Because now we don't have just one, but two rather big databases.

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=988322