Merge the camera import menus, so that we have only one list with cameras to choose from. All auto-detected cameras will have a suffix and they are separated by a line from the manually added cameras.
For example:

Nikon D50 (PTP Mode, auto-generated)
------------------------------------
Canon Powershot A80 (PTP Mode)
Hama Cardreader 8 in 1
------------------------------------
Add Camera...

I had to restructure the code a little bit, for example we fill the
solidUsmActionGroup and solidCameraActionGroup all the time, so that
duplicate entries may occur. This has been fixed (I hope), but for
manualCameraActionGroup it is still valid. Every time you add a camera,
the list is duplicated. I will fix that later.

Some polishing needs to be done, too.
Another thing to add is a way to determine duplicate entries and remove
the auto-detected ones (the manually added cameras are the primary
target and should always be displayed).
With the CameraNameHelper class this tasks should be easy, I will add
this later, too.

svn path=/trunk/extragear/graphics/digikam/; revision=1005839