Do not add duplicate camera actions.
This is a imple comparison method, based on some string operations.
Somehow auto-detection gives different mode names in some cases, so
simply comparing both modes is not working.
This method should be able to compare (at least with PTP mode ;-))
camera names and decide if they are the same. If so, we don't add the
auto-detected version and only present the manually added camera.

In the future we might want to work with setData() for the camera
actions. We could add a QMap<QString, QVariant> to it which has values
like

vendor
product
protocol
port

etc. This way the CameraNameHelper might become obsolete again, or at
least the comparison method.

svn path=/trunk/extragear/graphics/digikam/; revision=1005934