change order of filetype patterns: 1. All Images 2. All RAW files 3. the rest...
For me the filedialog is broken, can you confirm this (it was broken
before this commit, before anyone asks ;-)).
If I open showFoto and click on "Open" (not "Open Folder"), I can select
any filter, I will not see my files. The only thing that seems to work
is "All RAW files", but still when hitting "Ok" the photo is not loaded
into showFoto.

Can someone confirm this?

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=1006012