New utility: ManualRename
This utility will replace the old ManualRenameInput widget. The goal of
this utility is to have a more flexible parser that can be expanded
quite easily.

To create a new parser, inherit from Digikam::ManualRename::Parser and
overwrite the parse() method (and, if needed, slotTokenTriggered()).

A parser can register a button or menu in the client widget, right now
the only client widget is ManualRenameWidget, that behaves like the old
ManualRenameInput widget. These buttons and menus can be used to access
a token, so that the user doesn't have to type in the actual parse
string.

Every parser has one or more token objects. The token information can be
used to create the menu actions or the button. In the
ManualRenameWidget, this information is also used to create the tooltip
helper button.
So if you add a new parser, the button, menu and tooltip is
automatically created, to minimize coding efforts.

Some parsers might want to provide additional dialogs, to show / use
them overwrite the slotTokenTriggered() method and create the dialog in
there.

The parse information has to be provided to the parse() method with a
ParseInformation object.
If you create a new parser and need additional information (right now
filename, cameraName, dateTime and index is provided), add the extra
information to the ParseInformation class and make sure the client fills
in the correct values.

The layout of the ManualRenameWidget can be controlled with flags
(TokenInputStyle):
- None: No control buttons or toolbutton with menu is shown
- BigButtons: the widget uses big buttons below the KLineInput widget.
The column number can be set in the registerParsers(int columns) method.
(used in CameraUI)
- ToolButton: a small toolbutton is shown next to the KLineInput widget.
  This is normally useful if you don't have much space in the client.
(used in BQM)

The TokenInputStyle can be combined, if needed:
setTokenInputStyle(BigButtons|ToolButton)

In the future we might even remove the "Custom" rename widgets from the
camera import dialog, because it actually is less powerful and you are
not able to change the order of the added information.
What do you think? I guess with the quick-access buttons, the
ManualRenameWidget is quite user friendly now, so the custom widget can
be removed.

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=1012412