Complete re-write of the ManualRename utility parser. New feature:
Modifiers can be applied to tokens. A modifier can be used to change the
case of a string, to replace a portion of a parsed token etc.

Modifiers are not aware of the token parsers, which is actually good (no
need to add the modifiers in the parse code),
but also caused a total re-write of the parser code. Hope everything is
still working fine, please test this.

In the future there might be a control widget to add modifiers to a
marked token.
At the moment only one token can be assigned to a token, need to find
out how to allow more without breaking code again or increasing
complexity.

To use a modifier, just append it to a token.

EXAMPLE (no modifier):

parseString: [dir]_###_[file]_xyz
filename: DSC3245.jpg
dirname: testImages
result: testImages_001_DSC3245_xyz

WITH MODIFIER:

parseString: [dir]&_###_[file]%_xyz
result: TESTIMAGES_001_dsc3245_xyz

svn path=/trunk/extragear/graphics/digikam/; revision=1023587