When an album has been removed, set the search folder view to the current search again, otherwise we get crashes. A lot of album pointers in SearchFolderView still point to the currently deleted album, and since the album has been removed, we get a SegFault.
For example:
Save a search, remove it again. The line edit for editing the simple search
keywords is still activated and you can enter a new keyword for the deleted
search. Hit enter: boom.
This is the backtrace Michael has posted here.

The crashes / invalid reads reported by valgrind are something I don't
understand. But you are right, it can crash any FolderSearchView (GPS,
TimeLine, Search, FuzzySearch) etc. But how to reproduce, I can't say. It
sometimes happens, sometimes not.

Another problem:
Create a search in GPS, call it 'aaa'.
Create another search in Search, also call it 'aaa', you get a dialog that says
that a search with this name already exists.
Sure, behind the scenes all those searches are of one type, but for the normal
user it is confusing when it reads "The search already exists" and he simply
can't see it in the FolderView.

Marcel,
any idea how to fix this?

In the meantime I will try to figure out how to crash GPSSearch with a defined
"pattern".

CCBUG:209207

svn path=/trunk/extragear/graphics/digikam/; revision=1031092