Do not set the newSize to -1 of no rotation has been applied, but set the original size.
These labels are still wrong and (for me) not really needed. If you rotate an
image, it will show the newSize of the previewed image data, not the original
one.
This is actually wrong, because the size doesn't represent the real outcome.

I would suggest to remove those labels completely, but I guess this should be
discussed before ;-)

So what do you think?

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=1038521