Fix (hopefully at last) the FolderViews. Qt3 and Qt4 acts strange here, columns are either resized to the list view width (this is stupid) or it maximizes the column, but doesn't grow the column when the view is actually wider than the column.
We now calculate the column width our self, and it seems to finally work fine.
A too small view will show scrollbars, a wider view will switch to
LastColumn-ResizeMode.

To see the old "weird" behavior in Qt4, just fire up Kmail and resize the
folder view, the last column will stay where it is.
But this is necessary to have a scrollbar, because otherwise the column is
always resized.

I think this is a bug, because a normal QScrollArea acts not like that, but a
ListView somehow does.

CCBUG:141238

svn path=/trunk/extragear/graphics/digikam/; revision=1044060