Add Syntax Highlighting to AdvancedRename utility. Selecting tokens has been removed again, I guess nobody would have used it anyway. Since I had to embed an QTextEdit into a ProxyLineEdit and then into a QComboBox, it is possible that something is not working correctly anymore ;-)
Could you please check if the widget behaves normally?
Does it look normal?
I have size issues with it, see the FIXME comment in AdvancedRenameInput.
Any ideas how to solve this?

Do you know if the clear-button from KLineEdit is a single class somewhere, so
that I can embed it next to the QTextEdit again?
Right now you have to select the text manually to remove it.

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=1051584