Change the way the parser handles numbering. An internal counter is used to number the images, with an additional startIndex parameter used by CameraUI to set a different start than 1.
CameraUI increases the number / index while it has not been closed, which is a
good / normal behavior. So when you select three images, download them and
select different ones later on from the same camera device, the next index
will be 4.

Resetting this counter manually was not possible the way the rename parser
worked before.

Now the parser can be initialized with a ParseInformation object, that can have
a custom startIndex parameter.

BUG:215715

svn path=/trunk/extragear/graphics/digikam/; revision=1055194