Add experimental modifier: 'Unique'. I'm not totally sure how to handle this, maybe I need to restructure the parser in the future to hold information about all files that should be renamed. Right now it only renames one file after the other (which is actually a good thing, but may be problematic in some cases).
This modifier will allow to have unique values from a renaming option, by
adding a number prefix if necessary.
The problem at the moment: I can not reset the modifier for every renaming
option, so if you use it more than once, the counter just increases.

I will add a little demo video on my Youtube account to show how it works.

bdbrot,
what do you think? Is this enough for you?
I explicitly implemented it as a modifier so it can be attached to any renaming
option, not only the date field.

Andi

CCBUG:208920

svn path=/trunk/extragear/graphics/digikam/; revision=1055252