Prevent SearchTextbar from complaining about missing entries in idToTextMap.
This was caused by some really unhappy interactions of the AbstractAlbumModel and the AbstractCountingAlbumTreeView. The SearchTextBar complains if dataChanged is signaled without having received a rowsInserted for the changedAlbums. This happened because of this interaction chain:
1. the model signals rowsInserted.
2. the first receiver (before the search bar) is the counting tree view
3. this does some operations here that trigger recalculating the album count in the model based on the expansion state of the tree view
4. the recalculation of the album count is displayed in the item data, thus dataChanged is signaled
5. SearchTextBar receives the dataChanged for the album and complains...
6. After this complete chain, the signal rowsInserted is finally also sent to SearchTextBar, too late...

I've prevented this by using rowsAboutToBeInserted in the SearchTextBar. This works but I'm not sure if we got a more serious problem with this interaction. I would expect that the model ensures for every view that it receives an inserted before a changed and that's not valid here anymore. What do you think, Marcel?

CCMAIL: marcel.wiesweg@gmx.de

svn path=/trunk/extragear/graphics/digikam/; revision=1086357