* Prevent items from popping up again. * Added a TODO about a wrong assumption concerning filtering
Marcel, returning from a filter doesn't work eg. in the tag list of ImageDescEditTab. After filtering everything is collapsed. This is because the current album for this view isn't determined by the current album of the AlbumManager (see filterChanged). We must somehow save the complete state of the tree view before the filtering starts and restore it after filtering. Do you have an idea how to do this? I'm not sure what the best solution is.

CCMAIL: marcel.wiesweg@gmx.de

svn path=/trunk/extragear/graphics/digikam/; revision=1086574