When changing the rating, keep showing the overlays.
Changing the rating causes refiltering of the view, which causes
the layoutChanged() signal, and relayouting of the view.
In this case hiding the overlays is necessary, but after the view's relayout,
we can emit entered() again if the mouse is over an item.
Currently, this means the widgets are faded in again, which is a bit
annoying.

BUG: 228810

svn path=/trunk/extragear/graphics/digikam/; revision=1101296