Fix loading of date album on application start.
The problem was that DAlbums are created asynchronously. The sidebar tab was activated before the albums are loaded which resulted in setting 0x0 as current album because the tree view was not populated. To prevent this the DateFolderView now listens on signaleAllDAlbumsLoaded and re-sets the current album in the album manager if this signal occurs and the folder view is active.

BUG: 236002

svn path=/trunk/extragear/graphics/digikam/; revision=1124554