Avoid defining the content of libadvancedrename_SRCS twice.
But why do we do it like this anyway?
Why don't we use static libs for all the lib***_SRCS defines?
For example AdvancedRename is compiled three times, for the digiKam target, for
the tests target and for the themedesigner target.
This is true for a lot of other things, too.

Why don't we use static libs and just link them in the end?
Other candidates would be libdigikamgui, libsearchwindow etc...

Is there any reason why it is not done this way?

CCMAIL:digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=1133923