Add wrapper classes for two of the new tables:
TagProperties provides easy access to the properties of a tag.

ImageTagPair provides a similar interface for a single pair of an image
and a tag. Note that the tag need not necessarily be assigned to the image;
the new class allows to do that as well.
There is a method to find all tags with image-tag properties for a given image,
also available from the ImageInfo object.

CCMAIL: digikam-devel@kde.org

svn path=/branches/extragear/graphics/digikam/; revision=1146835