Now it's possible to draw the history in sidebar either as a list, or as a tree. It can be changed by the button on the versions widget.
CCMAIL: marcel.wiesweg@gmx.de

Marcel, would it be possible for getRelationCloud to only return the direct relations (eg. parent/child)? That way we wouldn't have to check whether the image is in the returning list already or not, otherwise it could have negative impact on performance on large versions collection. And it has to be checked now, because otherwise for one subversion it would show up twice in the view on two different tree levels, for second-level subversion (org->v1->sv1->ssv1) it would show up three times etc.


svn path=/branches/extragear/graphics/digikam/; revision=1159466