Add new class AdvancedRenameManager that holds information about files to be renamed. Before this class renaming was done by parsing one file after the other, so the renaming options and modifiers had no information about already parsed files and how many files there will be renamed in general.
For extension aware numbering it is important to know all files that should be
parsed.
In the future this class will maintain the whole parsing process. Right now it
is just a "hack" to get things done, but to integrate it better into digiKam,
a lot of code needs to be changed, especially in CameraUI and BQM.
I don't have too much time at the moment, so this is the result :-)

At least it works fine here.

This commit also fixes some other problems:

- BQM: when using the "Convert to..." options, file information could not be
  read, like camera name and metadata in general. This should be fixed now, BUT
  you are not able to modify the file extension in BQM (e.g. upper- or
  lowercase).

- Date information was sometimes not available in CameraUI.

BUG:246938

svn path=/trunk/extragear/graphics/digikam/; revision=1160586