Finally I found out why filtering by tag property didn't work.
When a new album is added or an existing is changed, setDynamicSortFilter()
will update the filtering for this index; but in our filtering approach,
which also displays children and parents of matches, a change may effect
the filtering of other indexes as well.
So we need to listen for changes and completely invalidateFilter() then.

Implement that with the help of a new signal in AlbumManager.


svn path=/branches/extragear/graphics/digikam/; revision=1164239