Polish UI of the AssignNameWidget.
Looks now much more like I want it to. There are still some rough
edges though (and missing functionality)

It would be desirable to have the popup boxes really pop out
and not be clipped to the scene, but this behavior is broken in Qt 4.6.2.

There are also two different popup boxes now, one from the line edit
and one from the combo box. There is room for improvement as well.

svn path=/branches/extragear/graphics/digikam/; revision=1181383