set the lastGlobalThumbSize, otherwise it is always zero and therefore a wrong label is set for the dtooltiptracker as well as the slider on album switch.
There is still some weird issue:
Everytime you set an thumbsize (e.g. 80) and switch to another album, the size
is always
size +=2
even if you use the maxsize of 256 (it becomes 258).

This only happens when the slider has been moved (a new thumbsize has been
defined), if you switch the album a second time, the size+=2 version is used.

svn path=/trunk/extragear/graphics/digikam/; revision=1198151