on delete trigger did a bit too much cleanup
according to:
http://dev.mysql.com/tech-resources/articles/hierarchical-data.html
only elemnts between `lft` and `rgt` need to be deleted, not the whole
hiearchy, also populate it again is not needed.

Maybe during migration or in some other particular case better handling
is required. These cases still need to be found and addressed ATM.

This possibly close
CCBUG: 272301