appliction disable "internal database"
When internal database has been disabled at compile time we don't know
how to start a server and widget should not show the option to use one.

Also for how cmake machinery work mysqldPath option is not empty but
contain "SERVERCMD_MYSQL-NOTFOUND", add a check for that.

When a digikam w/o mysqld support encounter a configuration file which
point to a "internal database" it threat the database as external and
will fail to connect to.

GUI:
BUG: 254533