Add root to Tree table and fix some related query
As explained in bug 276356 tree table need a root for sql to work well.
The startup query now insert an element with id = 0

During the cleanup I've decided to get rid of the confusing and mostly
redundant TagsTree table.

A bug found is that tags query had named parameter :tagID mentioned
multiple times in the query which is forbidden by QT implementation.

This patch still need testing and as such will be commited in sql2.0
branch but should arrive in master soon

TODO:
 - check the SQLite version
 - alter the c++ code where needed
 - DELETE in the tags tree is not managed by triggers anymore, will be
   managed by multiple query.
CCBUG: 276356