fix memory leak: DImgFilterGenerator objects are not deleted properly
We save the DImgFilterGenerator objects in the filterMap and coreGenerators, but we only cleanup coreGenerators in the destructor.
This results in a memory leak.
We can not simply delete the filterMap, because the DImgFilterGenerator pointers are shared and therefore the application will crash due to multiple delete calls.
Use a QSharedPointer to avoid the problem.

A new method addGenerator(const ImgFilterPtr& generator) has been added to the private class, so the code is still API compatible.