Rewrite DIO: Before starting a job, check for sidecar files which need to be handled alongside.
The check involves disk access and can take longer, therefore it needs to be done in a thread.
That means all the DIO methods no longer return a KIO::Job, instead the DIO class now listens
for the result signal. This means code cleanup in the dependent classes.
DIO class has become a singleton, the thread operations are done with an internal WorkerObject.