Implement training.
Please remove any old recognition.db before.
To retrain all images, there is a check box in the Advanced tab of the Scan Faces dialog.
Runs through, does not crash. Need to check in more depth that it does the right thing.
Before we expect decent results, we need to add lightning adjustment and face alignment to libkface.