Display the download file in the icon view for the selected items, like we did in digiKam 2.x. I still think this is not the best solution, but it seems to work right now.
There is a refresh problem with the view, that has always been there and is not introduced by this patch.
If you change rename settings or if you download images, the icon view is only refreshed when you move the mouse out of the view and in again.
I tried fixing this issue, but failed.
Maybe someone who knows this code can take a look at the issue?

Andi

CCBUG:314441