Improve digiKam startup time : ------------------------------
- remove splashscreen from ScanController. Always use a progress dialog instead in all cases.

- We will always require a progress dialog at first init to scan albums.
All details are given in dialog about which album is scanned and how many items still in the queue to process.

- The splashscreen always arrive after the progress dialog. Splash show final stage to init GUI

- When GUI is displayed, always start a delayed find new item process through progress manager. This is done in background.
Progress info is given in progress manager as well. While scan for new item is running in a separated thread,
user can start to work with GUI, even if all items are not yet catalogued in database.

BUGS: 320359
BUGS: 281959
FIXED-IN: 5.1.0