Support to migrate digiKam 4 settings/database in the first run wizard
This commit adds a new page to the FirstRun wizard that allows the user
to choose if he/she wants to do a migration of the digikam4 settings
(only if settings from digikam4 are detected). If the user selects to
create new settings, then the usual wizard continues, and if the user
selects to migrate then
~/.kde4/share/config/digikamrc
~/.kde4/share/config/digikam_tagsmanagerrc
~/.kde4/share/config/kipipluginsrc
~/.kde4/share/config/kipirc
~/.kde4/share/config/showfotorc
are copied to ~/.config/ .

Then all files inside ~/.kde4/share/apps/digikam/
are copied to ~/.local/share/apps/digikam .

And finally, the albumroots are updated in the database since
digiKam 5 doesn't interpret correctly values like
volumeid:?path=%2Fhome%2Fantonio%2FPictures and they need to be
url-decoded.