Replaced the minimum and maximum similarity selection boxes in fuzzy search, duplicates search and maintenance by a new, generic DIntRangeBox which makes the look, feel and behavious consistent in all places. Also, this makes redundant code obsolete and reduces the probability of bugs.
The new range box can be configured to an arbitrary integral minimum and maximum range
but is per default defined for values between 0 and 100. The range box enforces
total ordering of the upper and lower value and also allows an arbitrary value suffix
which is defined both for the upper and the lower value.

Also, I fixed some inconsistent uses of my email address in license headers.