fix cppcheck warning about non explicit constructor.
Maik, this fix is important, as depending of internal compiler rules, the PAlbumPath constructor used to handle the one argument
PAlbum* can be confused with the constructor taken an albumid instead, even if this one take 2 arguments.
The proff is all error generated by GCC when i introduced the explicit keyword on front of this one argument constructor. In fact,
GCC is lost to take the right constructor when when host PAlbumPath to the QHash table, epecially with [] operator.
I don't think the previous code is bugous as well, but depending of the compiler version the constructor binary code can be a garbage
and introduce crash in specific conditions.
In all cases, i think this code will be more robust in time.

CCMAIL: metzpinguin@gmail.com